Modernise Zygote.
