+++ tools/zygote/Dockerfile
@@ -1,8 +1,9 @@
 FROM julia:1.10.4
 WORKDIR /home/gradbench
-COPY tools/zygote/Project.toml .
-COPY tools/zygote/Manifest.toml .
-RUN julia --project=. -e 'import Pkg; Pkg.instantiate()'
-COPY tools/zygote/run.jl .
-ENTRYPOINT ["julia", "--project=/home/gradbench", "/home/gradbench/run.jl"]
+
+COPY tools/zygote /home/gradbench/tools/zygote
+
+RUN julia --project=tools/zygote -e 'import Pkg; Pkg.instantiate()'
+
+ENTRYPOINT ["julia", "--project=/home/gradbench/tools/zygote", "/home/gradbench/tools/zygote/run.jl"]
 LABEL org.opencontainers.image.source=https://github.com/gradbench/gradbench
