+++ tools/zygote/run.jl
@@ -1,18 +1,4 @@
-module hello
-
-import Zygote
-
-function square(x)
-  return x * x
-end
-
-function double(x)
-  z, = Zygote.gradient(square, x)
-  return z
-end
-
-end
-
+include("./hello.jl")
 import JSON
 
 function resolve(mod, name)
